requirements.txt
azure-functions
xgboost
pandas
scikit-learn
numpy
pyarrow